html/semantics/embedded-content-0/media-elements/preload_property_exists.html

<!doctype html>
<html>
 <head>
  <title>{audio,video}.preload - existence test</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
 </head>
 <body>
  <p><a href="http://dev.w3.org/html5/spec/Overview.html#dom-media-preload">spec reference</a></p>
  <audio id="audio">
  </audio>
  <video id="video">
  </video>
  <div id="log"></div>
  <script>
test(function() {
 assert_idl_attribute(
   document.getElementById("audio"),
   "preload",
   "audioElement.preload property exists");
}, "audio.preload - existence test");

test(function() {
 assert_idl_attribute(
   document.getElementById("video"),
   "preload",
   "videoElement.preload property exists");
}, "video.preload - existence test");
  </script>
 </body>
</html>
